converted the system module to use the action controllers
