+++ ACP3/Modules/ACP3/System/Controller/Admin/Maintenance/Index.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\System\Controller\Admin\Maintenance;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\System;
+
+/**
+ * Class Index
+ * @package ACP3\Modules\ACP3\System\Controller\Admin\Maintenance
+ */
+class Index extends Core\Modules\AdminController
+{
+    public function execute()
+    {
+        return;
+    }
+}
